:tractor: Separate submodules into individual files
Move the subparser and the Handler into single files to keep things
overseeable.
